He ido haciendo modificaciones al código para que funcione con Vitis HLS, sin embargo, no he podido comprobar la funcionalidad del mismo y garantizar la integridad. Sin embargo, parece que sí que la mantiene porque sólo hemos hecho cambios en la forma de la llamada.

+++ srcs/lib/simulink_block.h
@@ -1,1 +1,1 @@
-void myFunction(double x_ini[3], double Vsd, double Vsq, double iL, double u00[2], double outputVector[2]);
+void myFunction(float x_ini[3], float Vsd, float Vsq, float iL, float u00[2], float outputVector[2]);
